+++ StonksMobile/app/globals/styles/AppStyles.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+const appStyles = StyleSheet.create({
+    flexContainer: {
+        flex: 1,
+        width: "100%"
+    }
+});
+
+export default appStyles;
