+++ StonksMobile/app/screens/frontPage/Styles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    text: {
+        fontFamily: 'Poppins_400Regular',
+        fontSize: '48px',
+        color: '#ffffff',
+    }
+});
+
+export default styles;
